refactor auth service and controller
